requirements.txt
langchain
langchain-openai
langchain-pinecone
openai
pinecone-client
pypdf
streamlit
python-dotenv
tiktoken